refactor(admin): migrate stats tab to canonical /social/... path
