+++ apps/web/src/app/social/[platform]/[handle]/page.tsx
@@ -1,10 +1,33 @@
-import { redirect } from "next/navigation";
+import { notFound, redirect } from "next/navigation";
+import SocialAccountProfilePage from "@/components/admin/SocialAccountProfilePage";
+import {
+  SOCIAL_ACCOUNT_PROFILE_PLATFORMS,
+  type SocialPlatformSlug,
+} from "@/lib/admin/social-account-profile";
+import { normalizeSocialAccountProfileHandle } from "@/lib/admin/show-admin-routes";
 
 type PageProps = {
   params: Promise<{ platform: string; handle: string }>;
 };
 
-export default async function LegacySocialAccountProfileStatsPage({ params }: PageProps) {
+const isValidHandle = (value: string): boolean => /^[a-z0-9._-]{1,64}$/i.test(value);
+
+export default async function SocialAccountProfileStatsPage({ params }: PageProps) {
   const resolved = await params;
-  redirect(`/admin/social/${encodeURIComponent(resolved.platform)}/${encodeURIComponent(resolved.handle)}`);
+  const platform = resolved.platform.trim().toLowerCase();
+  const rawHandle = resolved.handle.trim().replace(/^@+/, "").toLowerCase();
+  if (!SOCIAL_ACCOUNT_PROFILE_PLATFORMS.includes(platform as (typeof SOCIAL_ACCOUNT_PROFILE_PLATFORMS)[number])) {
+    notFound();
+  }
+  if (!isValidHandle(rawHandle)) {
+    notFound();
+  }
+  const handle = normalizeSocialAccountProfileHandle(rawHandle);
+  if (!handle) {
+    notFound();
+  }
+  if (rawHandle !== handle) {
+    redirect(`/social/${encodeURIComponent(platform)}/${encodeURIComponent(handle)}`);
+  }
+  return <SocialAccountProfilePage platform={platform as SocialPlatformSlug} handle={handle} activeTab="stats" />;
 }
